fix no html html

+++ index.js
@@ -31,12 +31,17 @@ function copyFiles(outPath) {
 }
 
 module.exports = function(bundler) {
-  bundler.on('bundled', bund => {
-    const bundleDir = path.dirname(bund.name);
-    const htmlPath = path.resolve(bundleDir, 'index.html');
-    const data = fse.readFileSync(htmlPath, { encoding: 'utf-8' });
-    fse.removeSync(htmlPath);
-    changeHtml(htmlPath, data);
-    copyFiles(bundleDir);
-  });
+  if (process.env.changeFile != 'false') {
+    bundler.on('bundled', bund => {
+      const bundleDir = path.dirname(bund.name);
+      const htmlPath = path.resolve(bundleDir, 'index.html');
+      const ishaveHtml = fse.existsSync(htmlPath);
+      if (ishaveHtml) {
+        const data = fse.readFileSync(htmlPath, { encoding: 'utf-8' });
+        fse.removeSync(htmlPath);
+        changeHtml(htmlPath, data);
+      }
+      copyFiles(bundleDir);
+    });
+  }
 };
